--- UniespTech-main/br/uniesp/Model/Aluno.java
@@ -1,17 +1,0 @@
-package br.uniesp.Model;
-
-public class Aluno {
-    private String nome;
-    private String cpf;
-
-    public Aluno(String nome, String cpf) {
-        this.nome = nome;
-        this.cpf = cpf;
-    }
-    public String getNome() {
-        return nome;
-    }
-    public String getCpf() {
-        return cpf;
-    }
-}
