--- UniespTech-main/br/uniesp/SistemaUniesp.java
@@ -1,13 +1,0 @@
-package br.uniesp;
-
-import br.uniesp.Controller.AlunoController;
-
-public class SistemaUniesp {
-
-    public static void main(String[] args) {
-
-        AlunoController controller = new AlunoController();
-        controller.iniciar();
-
-    }
-}
