+++ UniespTech-main/src/SistemaUniesp.java
@@ -1,0 +1,13 @@
+package src;
+
+import Controller.AlunoController;
+
+public class SistemaUniesp {
+
+    public static void main(String[] args) {
+
+        AlunoController controller = new AlunoController();
+        controller.iniciar();
+
+    }
+}
