+++ UniespTech-main/src/main/java/Model/Aluno.java
@@ -1,0 +1,17 @@
+package Model;
+
+public class Aluno {
+    private String nome;
+    private String cpf;
+
+    public Aluno(String nome, String cpf) {
+        this.nome = nome;
+        this.cpf = cpf;
+    }
+    public String getNome() {
+        return nome;
+    }
+    public String getCpf() {
+        return cpf;
+    }
+}
