+++ UniespTech-main/src/main/java/Repository/AlunoRepository.java
@@ -1,0 +1,18 @@
+package Repository;
+
+import java.util.ArrayList;
+import Model.Aluno;
+
+public class AlunoRepository {
+    private ArrayList<Aluno> alunos =  new ArrayList<>();
+    public void cadastrar (Aluno aluno) {
+        alunos.add(aluno);
+    }
+
+    public ArrayList<Aluno> listar() {
+        return alunos;
+    }
+    public void deletartodos() {
+        alunos.clear();
+    }
+}
